plugin docstring and renamed _ensemble_data.py to _ensemble_well_analysis_data.py
